+++ ifood/cadastro/src/main/java/github/com/marcoant007/ifood/repository/LocalizacaoRepository.java
@@ -6,6 +6,6 @@ import github.com.marcoant007.ifood.models.Localizacao;
 import io.quarkus.hibernate.orm.panache.PanacheRepositoryBase;
 
 @ApplicationScoped
-public class LocalizacaoRepository implements PanacheRepositoryBase<Localizacao, Integer> {
+public class LocalizacaoRepository implements PanacheRepositoryBase<Localizacao, Long> {
     
 }
